added repositories and service-classes

+++ src/main/java/de/bauersoft/data/repositories/day/DayRepository.java
@@ -3,5 +3,5 @@ package de.bauersoft.data.repositories.day;
 import de.bauersoft.data.entities.Day;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DayRepository extends JpaRepository<Day, Integer> {
+public interface DayRepository extends JpaRepository<Day, Long> {
 }
