+++ src/main/java/de/bauersoft/data/repositories/week/WeekRepository.java
@@ -3,5 +3,5 @@ package de.bauersoft.data.repositories.week;
 import de.bauersoft.data.entities.Week;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WeekRepository extends JpaRepository<Week, Integer> {
+public interface WeekRepository extends JpaRepository<Week, Long> {
 }
